--- a/miden-crypto/src/hash/algebraic_sponge/rescue/arch/mod.rs
+++ b/miden-crypto/src/hash/algebraic_sponge/rescue/arch/mod.rs
@@ -43,12 +43,15 @@ pub mod optimized {
     }
 }
 
-#[cfg(target_feature = "avx2")]
+#[cfg(all(target_feature = "avx2", not(target_feature = "avx512f")))]
 mod x86_64_avx2;
 
-#[cfg(target_feature = "avx2")]
+#[cfg(target_feature = "avx512f")]
+mod x86_64_avx512;
+
+#[cfg(target_feature = "avx512f")]
 pub mod optimized {
-    use super::x86_64_avx2::{apply_inv_sbox, apply_sbox};
+    use super::x86_64_avx512::{apply_ext_round, apply_inv_sbox, apply_sbox};
     use crate::{
         Felt,
         hash::algebraic_sponge::rescue::{STATE_WIDTH, add_constants},
@@ -61,7 +64,7 @@ pub mod optimized {
     ) -> bool {
         add_constants(state, ark);
         unsafe {
-            apply_sbox(std::mem::transmute(state));
+            apply_sbox(core::mem::transmute(state));
         }
         true
     }
@@ -73,13 +76,70 @@ pub mod optimized {
     ) -> bool {
         add_constants(state, ark);
         unsafe {
-            apply_inv_sbox(std::mem::transmute(state));
+            apply_inv_sbox(core::mem::transmute(state));
+        }
+        true
+    }
+
+    #[inline(always)]
+    pub fn add_constants_and_apply_ext_round(
+        state: &mut [Felt; STATE_WIDTH],
+        ark: &[Felt; STATE_WIDTH],
+    ) -> bool {
+        add_constants(state, ark);
+        unsafe {
+            apply_ext_round(core::mem::transmute(state));
         }
         true
     }
 }
 
-#[cfg(not(any(target_feature = "avx2", target_feature = "sve")))]
+#[cfg(all(target_feature = "avx2", not(target_feature = "avx512f")))]
+pub mod optimized {
+    use super::x86_64_avx2::{apply_ext_round, apply_inv_sbox, apply_sbox};
+    use crate::{
+        Felt,
+        hash::algebraic_sponge::rescue::{STATE_WIDTH, add_constants},
+    };
+
+    #[inline(always)]
+    pub fn add_constants_and_apply_sbox(
+        state: &mut [Felt; STATE_WIDTH],
+        ark: &[Felt; STATE_WIDTH],
+    ) -> bool {
+        add_constants(state, ark);
+        unsafe {
+            apply_sbox(core::mem::transmute(state));
+        }
+        true
+    }
+
+    #[inline(always)]
+    pub fn add_constants_and_apply_inv_sbox(
+        state: &mut [Felt; STATE_WIDTH],
+        ark: &[Felt; STATE_WIDTH],
+    ) -> bool {
+        add_constants(state, ark);
+        unsafe {
+            apply_inv_sbox(core::mem::transmute(state));
+        }
+        true
+    }
+
+    #[inline(always)]
+    pub fn add_constants_and_apply_ext_round(
+        state: &mut [Felt; STATE_WIDTH],
+        ark: &[Felt; STATE_WIDTH],
+    ) -> bool {
+        add_constants(state, ark);
+        unsafe {
+            apply_ext_round(core::mem::transmute(state));
+        }
+        true
+    }
+}
+
+#[cfg(not(any(target_feature = "avx2", target_feature = "avx512f", target_feature = "sve")))]
 pub mod optimized {
     use crate::{Felt, hash::algebraic_sponge::rescue::STATE_WIDTH};
 
@@ -93,6 +153,14 @@ pub mod optimized {
 
     #[inline(always)]
     pub fn add_constants_and_apply_inv_sbox(
+        _state: &mut [Felt; STATE_WIDTH],
+        _ark: &[Felt; STATE_WIDTH],
+    ) -> bool {
+        false
+    }
+
+    #[inline(always)]
+    pub fn add_constants_and_apply_ext_round(
         _state: &mut [Felt; STATE_WIDTH],
         _ark: &[Felt; STATE_WIDTH],
     ) -> bool {