--- a/miden-crypto/src/hash/algebraic_sponge/rescue/rpx/mod.rs
+++ b/miden-crypto/src/hash/algebraic_sponge/rescue/rpx/mod.rs
@@ -1,8 +1,8 @@
 use super::{
     ARK1, ARK2, CAPACITY_RANGE, CubeExtension, DIGEST_RANGE, ElementHasher, Felt, FieldElement,
     Hasher, MDS, NUM_ROUNDS, RATE_RANGE, Range, STATE_WIDTH, Word, add_constants,
-    add_constants_and_apply_inv_sbox, add_constants_and_apply_sbox, apply_inv_sbox, apply_mds,
-    apply_sbox,
+    add_constants_and_apply_ext_round, add_constants_and_apply_inv_sbox,
+    add_constants_and_apply_sbox, apply_inv_sbox, apply_mds, apply_sbox,
 };
 #[cfg(test)]
 use super::{StarkField, ZERO};
@@ -178,9 +178,8 @@ impl Rpx256 {
         }
     }
 
-    /// (E) round function.
     #[inline(always)]
-    pub fn apply_ext_round(state: &mut [Felt; STATE_WIDTH], round: usize) {
+    fn apply_ext_round_ref(state: &mut [Felt; STATE_WIDTH], round: usize) {
         // add constants
         add_constants(state, &ARK1[round]);
 
@@ -197,6 +196,14 @@ impl Rpx256 {
         *state = CubicExtElement::slice_as_base_elements(&arr_ext)
             .try_into()
             .expect("shouldn't fail");
+    }
+
+    /// (E) round function.
+    #[inline(always)]
+    pub fn apply_ext_round(state: &mut [Felt; STATE_WIDTH], round: usize) {
+        if !add_constants_and_apply_ext_round(state, &ARK1[round]) {
+            Self::apply_ext_round_ref(state, round);
+        }
     }
 
     /// (M) round function.