This improves performance of RPO and RPX hash functions by leveraging **AVX512** on compatible x86_64 hardware.  

Previously, the RPX E-round was not vectorized, so only the S-box benefited from SIMD. 
With this PR, both **AVX2** and **AVX512** paths now have fully accelerated E-rounds.

Benchmarks were run on a **Zen 5 (AMD EPYC 9575F, 16-core slice)** server CPU.

### RPX256 Benchmarks (with and without e-round optimization)

| Mode                | AVX2  | AVX512  | AVX2 + e-round. | AVX512 + e-round. |
|---------------------|-------------|---------------|-------------|---------------|
| 2-to-1 (cached)     | 2.85 µs     | 2.29 µs       | 2.49 µs     | 2.01 µs       |
| 2-to-1 (random)     | 2.85 µs     | 2.28 µs       | 2.50 µs     | 2.05 µs       |
| Sequential (cached) | 36.8 µs     | 31.8 µs       | 32.0 µs     | 27.7 µs       |
| Sequential (random) | 36.8 µs     | 31.0 µs       | 32.2 µs     | 28.1 µs       |

---

### RPO256 Benchmarks (S-box only)

| Mode                | AVX2  | AVX512. |
|---------------------|-------------|---------------|
| 2-to-1 (cached)     | 4.22 µs     | 3.89 µs  |
| 2-to-1 (random)     | 4.24 µs     | 3.93 µs     |
| Sequential (cached) | 57.0 µs    | 51.0 µs       |
| Sequential (random) | 56.3 µs    | 51.7 µs       |

### Highlights

- AVX512 (scalar) is already **~10–20% faster** than AVX2 scalar.  
- Accelerated E-round gives an additional **12–14% boost**.  
- Net result: **~30% faster RPX256 hashing overall** (`2.85 µs → 2.01 µs` for 2-to-1 cached).  
- RPO256 also benefits modestly (up to **11% faster sequential**).  

## Checklist before requesting a review
- Repo forked and branch created from `next` according to naming convention.
- Commit messages and codestyle follow [conventions](./CONTRIBUTING.md).
- Relevant issues are linked in the PR description.
- Tests added for new functionality.
- Documentation/comments updated according to changes.